requirements.txt
absl-py==2.0.0
affine==2.4.0
attrs==23.2.0
blinker==1.8.1
Brotli 
certifi
charset-normalizer
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
filelock
Flask==3.0.2
Flask-Cors==4.0.0
fsspec==2024.3.1
# GDAL==3.6.2
geoserver-rest==2.7.0
gmpy2 
google-auth-oauthlib==1.1.0
greenlet==3.0.3
grpcio==1.59.3
idna 
intel-openmp==2021.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
Markdown==3.5.1
MarkupSafe
mkl==2021.4.0
ml_collections
mpmath 
networkx 
nibabel==5.2.1
# numpy
oauthlib==3.2.2
opencv-python
pandas
pillow 
protobuf==4.23.4
pyparsing==3.1.2
PySocks 
PyYAML 
rasterio==1.3.10
requests 
requests-oauthlib==1.3.1
safetensors==0.4.2
snuggs==1.4.7
sympy 
scikit-learn==1.3.0
tbb==2021.12.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
thop==0.1.1.post2209072238
tokenizers==0.15.2
torch==2.3.0
torch-tb-profiler==0.4.3
torchaudio==2.3.0
torchsummary==1.5.1
torchvision==0.18.0
tqdm==4.66.2
typing_extensions
urllib3
Werkzeug==3.0.1
win-inet-pton
xmltodict==0.13.0
zope.event==5.0
